// api/backend/models/Error.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type Error = {
    data?: Record<string, any> | null;
    error: {
        status?: number;
        name?: string;
        message?: string;
        details?: Record<string, any>;
    };
};