// api/backend/models/Users_Permissions_Role.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type Users_Permissions_Role = {
    id?: number;
    name?: string;
    description?: string;
    type?: string;
    createdAt?: string;
    updatedAt?: string;
};